src/test/javascript/spec/app/entities/doc-flux/doc-flux.service.spec.ts
import { TestBed, getTestBed } from '@angular/core/testing';
import { HttpClientTestingModule, HttpTestingController } from '@angular/common/http/testing';
import { DocFluxService } from 'app/entities/doc-flux/doc-flux.service';
import { IDocFlux, DocFlux } from 'app/shared/model/doc-flux.model';

describe('Service Tests', () => {
  describe('DocFlux Service', () => {
    let injector: TestBed;
    let service: DocFluxService;
    let httpMock: HttpTestingController;
    let elemDefault: IDocFlux;
    let expectedResult: IDocFlux | IDocFlux[] | boolean | null;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule]
      });
      expectedResult = null;
      injector = getTestBed();
      service = injector.get(DocFluxService);
      httpMock = injector.get(HttpTestingController);

      elemDefault = new DocFlux(0, 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA', 'AAAAAAA');
    });

    describe('Service methods', () => {
      it('should find an element', () => {
        const returnedFromService = Object.assign({}, elemDefault);

        service.find(123).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(elemDefault);
      });

      it('should create a DocFlux', () => {
        const returnedFromService = Object.assign(
          {
            id: 0
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.create(new DocFlux()).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'POST' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should update a DocFlux', () => {
        const returnedFromService = Object.assign(
          {
            idEdition: 'BBBBBB',
            libelle: 'BBBBBB',
            modele: 'BBBBBB',
            famille: 'BBBBBB',
            version: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.update(expected).subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'PUT' });
        req.flush(returnedFromService);
        expect(expectedResult).toMatchObject(expected);
      });

      it('should return a list of DocFlux', () => {
        const returnedFromService = Object.assign(
          {
            idEdition: 'BBBBBB',
            libelle: 'BBBBBB',
            modele: 'BBBBBB',
            famille: 'BBBBBB',
            version: 'BBBBBB'
          },
          elemDefault
        );

        const expected = Object.assign({}, returnedFromService);

        service.query().subscribe(resp => (expectedResult = resp.body));

        const req = httpMock.expectOne({ method: 'GET' });
        req.flush([returnedFromService]);
        httpMock.verify();
        expect(expectedResult).toContainEqual(expected);
      });

      it('should delete a DocFlux', () => {
        service.delete(123).subscribe(resp => (expectedResult = resp.ok));

        const req = httpMock.expectOne({ method: 'DELETE' });
        req.flush({ status: 200 });
        expect(expectedResult);
      });
    });

    afterEach(() => {
      httpMock.verify();
    });
  });
});
